Add new shortcuts for indent navigation and line number memory
Adds new functionality to navigate up the current indent level(Ctrl-I)
and a shortcut to navigate back(Alt-K). There is also a shortcut(Alt-M)
to store the current line number in the stack for the current script.

This could be further enhanced with a navigate forward shortcut and
instead of pushing and popping the stack, using a position value that is
incremented or decremented.